refactor(cli): extract bin-only command branches into core/cli/bin-commands.ts
bin/prjct.ts main() carried ~550 lines of args[0]===X handlers; it now
delegates to runBinCommand(args, ctx) and keeps only startup concerns
(flag pre-parse, fast paths, setup check, auto-update). 1082 -> 536
lines. Same dynamic-import discipline (the daemon fast path pays for
nothing new). Moving the code under core/tsconfig.json surfaced a
latent type error: daemon start --port/--no-http were parsed but
startDaemon never accepted them — dead flags, dropped.
hook-dispatcher-parity now reads the extracted module and asserts the
manifest-derived command gate instead of source literals.
