+++ bin/prjct.ts
@@ -437,555 +437,9 @@ async function main(): Promise<void> {
     return () => {}
   }
 
-  if (args[0] === 'daemon') {
-    const subcommand = args[1] || 'status'
-
-    if (subcommand === 'start') {
-      const { isDaemonRunning, spawnDaemon } = await import('../core/daemon/client')
-
-      if (await isDaemonRunning()) {
-        console.log('Daemon is already running.')
-        process.exitCode = 0
-      } else {
-        const foreground = args.includes('--foreground') || args.includes('-f')
-
-        if (foreground) {
-          const { startDaemon } = await import('../core/daemon/daemon')
-          const port =
-            parseInt(args.find((a) => a.startsWith('--port='))?.split('=')[1] || '', 10) ||
-            undefined
-          const noHttp = args.includes('--no-http')
-          await startDaemon({ port, noHttp, foreground: true })
-        } else {
-          const started = await spawnDaemon()
-          if (started) {
-            console.log('Daemon started.')
-          } else {
-            console.error('Failed to start daemon.')
-            process.exitCode = 1
-          }
-        }
-      }
-    } else if (subcommand === 'stop') {
-      const { isDaemonRunning, stopDaemon } = await import('../core/daemon/client')
-
-      if (await isDaemonRunning()) {
-        const stopped = await stopDaemon()
-        console.log(stopped ? 'Daemon stopped.' : 'Failed to stop daemon.')
-        process.exitCode = stopped ? 0 : 1
-      } else {
-        console.log('Daemon is not running.')
-        process.exitCode = 0
-      }
-    } else if (subcommand === 'status') {
-      const { getDaemonStatus } = await import('../core/daemon/client')
-      const status = await getDaemonStatus()
-
-      if (status.running) {
-        const uptime = status.uptime ? Math.round(status.uptime / 1000) : 0
-        const stale = (status as Record<string, unknown>).stale
-        console.log(`Daemon running (PID ${status.pid})${stale ? ' ⚠ STALE' : ''}`)
-        console.log(`  Uptime: ${uptime}s`)
-        console.log(`  Commands served: ${status.commandsServed ?? 0}`)
-        if (status.lastActivity) {
-          console.log(`  Last activity: ${status.lastActivity}`)
-        }
-        if (stale) {
-          console.log(
-            `  ${chalk.yellow('⚠ Code changed since daemon started. Run: prjct restart')}`
-          )
-        }
-      } else {
-        console.log('Daemon is not running.')
-      }
-      process.exitCode = 0
-    } else if (subcommand === 'restart') {
-      const { isDaemonRunning, stopDaemon, forceKillDaemon, spawnDaemon } = await import(
-        '../core/daemon/client'
-      )
-
-      if (await isDaemonRunning()) {
-        const stopped = await stopDaemon()
-        if (!stopped) forceKillDaemon()
-        await new Promise((resolve) => setTimeout(resolve, 300))
-      } else {
-        forceKillDaemon()
-      }
-
-      const started = await spawnDaemon()
-      if (started) {
-        console.log('Daemon restarted.')
-        process.exitCode = 0
-      } else {
-        console.error('Failed to restart daemon.')
-        process.exitCode = 1
-      }
-    } else if (subcommand === 'logs') {
-      const fs = await import('node:fs')
-      const { DAEMON_PATHS } = await import('../core/daemon/protocol')
-      const logPath = DAEMON_PATHS.log()
-
-      if (!fs.existsSync(logPath)) {
-        console.error(`No daemon log at ${logPath}. Start the daemon first.`)
-        process.exitCode = 1
-      } else {
-        const follow = args.includes('--follow') || args.includes('-f')
-        const all = args.includes('--all')
-        const linesArg =
-          args.find((a) => a.startsWith('--lines='))?.split('=')[1] ||
-          (args.includes('-n') ? args[args.indexOf('-n') + 1] : undefined)
-        const lines = linesArg ? parseInt(linesArg, 10) : 50
-
-        if (follow) {
-          const { spawn } = await import('node:child_process')
-          const child = spawn('tail', ['-n', String(lines), '-f', logPath], { stdio: 'inherit' })
-          process.on('SIGINT', () => child.kill('SIGINT'))
-          await new Promise<void>((resolve) => child.on('exit', () => resolve()))
-        } else if (all) {
-          process.stdout.write(fs.readFileSync(logPath, 'utf-8'))
-        } else {
-          const content = fs.readFileSync(logPath, 'utf-8')
-          const allLines = content.split('\n')
-          const tail = allLines.slice(-Math.max(1, lines))
-          process.stdout.write(tail.join('\n'))
-          if (!content.endsWith('\n')) process.stdout.write('\n')
-        }
-        process.exitCode = 0
-      }
-    } else {
-      console.error(
-        `Unknown daemon command: ${subcommand}. Use: start, stop, restart, status, logs`
-      )
-      process.exitCode = 1
-    }
-  } else if (args[0] === 'stop') {
-    // Top-level shortcut: prjct stop → kill daemon (with force-kill fallback)
-    const { isDaemonRunning, stopDaemon, forceKillDaemon } = await import('../core/daemon/client')
-    const force = args.includes('--force') || args.includes('-f')
-
-    if (force) {
-      const killed = forceKillDaemon()
-      console.log(killed ? 'Daemon force-killed.' : 'Daemon is not running.')
-      process.exitCode = 0
-    } else if (await isDaemonRunning()) {
-      const stopped = await stopDaemon()
-      if (stopped) {
-        console.log('Daemon stopped.')
-        process.exitCode = 0
-      } else {
-        // Graceful stop failed — force kill
-        console.log('Graceful stop failed, force-killing...')
-        forceKillDaemon()
-        console.log('Daemon force-killed.')
-        process.exitCode = 0
-      }
-    } else {
-      // Check for stale files even if daemon not responding
-      forceKillDaemon()
-      console.log('Daemon is not running (cleaned up stale files).')
-      process.exitCode = 0
-    }
-  } else if (args[0] === 'restart') {
-    // Top-level shortcut: prjct restart → stop + start daemon
-    const { isDaemonRunning, stopDaemon, forceKillDaemon, spawnDaemon } = await import(
-      '../core/daemon/client'
-    )
-
-    // Stop first (graceful → force)
-    if (await isDaemonRunning()) {
-      const stopped = await stopDaemon()
-      if (!stopped) {
-        forceKillDaemon()
-      }
-      // Wait for process to fully exit
-      await new Promise((resolve) => setTimeout(resolve, 300))
-    } else {
-      // Clean up any stale files
-      forceKillDaemon()
-    }
-
-    // Start fresh
-    const started = await spawnDaemon()
-    if (started) {
-      console.log('Daemon restarted.')
-      process.exitCode = 0
-    } else {
-      console.error('Failed to restart daemon.')
-      process.exitCode = 1
-    }
-  } else if (args[0] === 'start' || args[0] === 'setup') {
-    const { runStart } = await import('../core/cli/start')
-    await runStart()
-  } else if (args[0] === 'context') {
-    const projectPath = process.cwd()
-    const projectId = await configManager.getProjectId(projectPath)
-
-    if (!projectId) {
-      console.error('No prjct project found. Run "prjct init" first.')
-      process.exitCode = 1
-    } else {
-      const done = await trackSession('context')
-      // Strip --md and --json flags before passing to context tools
-      const contextArgs = args.slice(1).filter((a) => a !== '--md' && a !== '--json')
-      const mdMode = args.includes('--md')
-
-      if (contextArgs.length === 0) {
-        // No subcommand: return project context
-        const { ContextCommands } = await import('../core/commands/context')
-        const contextCmds = new ContextCommands()
-        const result = await contextCmds.context(null, projectPath, { md: mdMode })
-        process.exitCode = result.success ? 0 : 1
-      } else {
-        const { runContextTool } = await import('../core/tools/context')
-        const result = await runContextTool(contextArgs, projectId, projectPath)
-        console.log(JSON.stringify(result, null, 2))
-        process.exitCode = result.tool === 'error' ? 1 : 0
-      }
-      done()
-    }
-  } else if (args[0] === 'hooks') {
-    const done = await trackSession('hooks')
-    const { hooksService } = await import('../core/services/hooks-service')
-    const subcommand = args[1] || 'status'
-    const exitCode = await hooksService.run(process.cwd(), subcommand)
-    process.exitCode = exitCode
-    done()
-  } else if (args[0] === 'seed') {
-    // `prjct seed add|remove|list|suggest [args]` — manage active packs.
-    const sub = args[1] ?? 'list'
-    const rest = args
-      .slice(2)
-      .filter((a) => !a.startsWith('-'))
-      .join(',')
-    const mdMode = args.includes('--md')
-    const { SeedCommands } = await import('../core/commands/seed')
-    const cmd = new SeedCommands()
-    let result: { success: boolean; error?: string } = { success: false, error: 'unknown' }
-    if (sub === 'add') result = await cmd.add(rest || null, process.cwd(), { md: mdMode })
-    else if (sub === 'remove')
-      result = await cmd.remove(rest || null, process.cwd(), { md: mdMode })
-    else if (sub === 'list') result = await cmd.list(null, process.cwd(), { md: mdMode })
-    else if (sub === 'suggest') result = await cmd.suggest(null, process.cwd(), { md: mdMode })
-    else {
-      console.error(`Unknown seed subcommand: ${sub}. Use: add, remove, list, suggest.`)
-    }
-    process.exitCode = result.success ? 0 : 1
-  } else if (args[0] === 'context-save') {
-    // `prjct context-save [title] [--notes "..."]` — checkpoint working
-    // state for resume in another session / branch / workspace.
-    const title = args.slice(1).find((a) => !a.startsWith('-')) ?? null
-    const notesIdx = args.indexOf('--notes')
-    const notes = notesIdx >= 0 ? args[notesIdx + 1] : undefined
-    const mdMode = args.includes('--md')
-    const { ContextCheckpointCommands } = await import('../core/commands/context-checkpoint')
-    const cmd = new ContextCheckpointCommands()
-    const result = await cmd.save(title, process.cwd(), { md: mdMode, notes })
-    process.exitCode = result.success ? 0 : 1
-  } else if (args[0] === 'context-restore') {
-    // `prjct context-restore [<file>] [--list] [--file <name>]` —
-    // emit the most recent checkpoint (or named one).
-    const list = args.includes('--list')
-    const fileIdx = args.indexOf('--file')
-    const file = fileIdx >= 0 ? args[fileIdx + 1] : undefined
-    const positional = args.slice(1).find((a) => !a.startsWith('-')) ?? null
-    const mdMode = args.includes('--md')
-    const { ContextCheckpointCommands } = await import('../core/commands/context-checkpoint')
-    const cmd = new ContextCheckpointCommands()
-    const result = await cmd.restore(positional, process.cwd(), { md: mdMode, list, file })
-    process.exitCode = result.success ? 0 : 1
-  } else if (args[0] === 'health') {
-    // `prjct health [--md]` — composite quality dashboard. Wraps
-    // typecheck / lint / tests / knip and reports a weighted score.
-    const mdMode = args.includes('--md')
-    const { HealthCommands } = await import('../core/commands/health')
-    const cmd = new HealthCommands()
-    const result = await cmd.health(null, process.cwd(), { md: mdMode })
-    process.exitCode = result.success ? 0 : 1
-  } else if (args[0] === 'retro') {
-    // `prjct retro [window]` — gstack-style weekly engineering retro.
-    // Window defaults to 7d; accepts NNh / NNd up to 365d.
-    const windowArg = args.slice(1).find((a) => !a.startsWith('-')) ?? null
-    const mdMode = args.includes('--md')
-    const { RetroCommands } = await import('../core/commands/retro')
-    const cmd = new RetroCommands()
-    const result = await cmd.retro(windowArg, process.cwd(), { md: mdMode })
-    process.exitCode = result.success ? 0 : 1
-  } else if (args[0] === 'skill-adherence') {
-    // `prjct skill-adherence [window]` — harness #16 QA surface: how
-    // often captured project knowledge went unreferenced, and how much
-    // got resolved. Read-only; window defaults to 7d.
-    const windowArg = args.slice(1).find((a) => !a.startsWith('-')) ?? null
-    const mdMode = args.includes('--md')
-    const { SkillAdherenceCommands } = await import('../core/commands/skill-adherence')
-    const cmd = new SkillAdherenceCommands()
-    const result = await cmd.skillAdherence(windowArg, process.cwd(), { md: mdMode })
-    process.exitCode = result.success ? 0 : 1
-  } else if (args[0] === 'review-risk') {
-    // `prjct review-risk [--md]` — advisory size/delivery-geometry
-    // signal (#18/19/20). Read-only; never gates or mutates git.
-    const mdMode = args.includes('--md')
-    const { ReviewRiskCommands } = await import('../core/commands/review-risk')
-    const cmd = new ReviewRiskCommands()
-    const result = await cmd.reviewRisk(null, process.cwd(), { md: mdMode })
-    process.exitCode = result.success ? 0 : 1
-  } else if (args[0] === 'prefs') {
-    // `prjct prefs list|get|check|set|clear [...]` — gstack-inspired
-    // per-project AskUserQuestion preferences. The `check` subcommand
-    // emits a one-line ASK_NORMALLY|AUTO_DECIDE|NEVER_ASK so a skill
-    // preamble can branch on its output.
-    const sub = args[1] ?? 'list'
-    const positional = args.slice(2).filter((a) => !a.startsWith('-'))
-    const reasonIdx = args.indexOf('--reason')
-    const reason = reasonIdx >= 0 ? args[reasonIdx + 1] : undefined
-    const mdMode = args.includes('--md')
-    const { PreferencesCommands } = await import('../core/commands/preferences')
-    const cmd = new PreferencesCommands()
-    const result = await cmd.prefs([sub, ...positional], process.cwd(), { md: mdMode, reason })
-    process.exitCode = result.success ? 0 : 1
-  } else if (args[0] === 'install') {
-    // `prjct install` is a convenience alias for `prjct claude install`.
-    const { InstallCommands } = await import('../core/commands/install')
-    const cmd = new InstallCommands()
-    const mdMode = args.includes('--md')
-    const result = await cmd.install(null, process.cwd(), { md: mdMode })
-    process.exitCode = result.success ? 0 : 1
-  } else if (args[0] === 'claude') {
-    // `prjct claude install|uninstall|status` — manage Claude Code hooks in ~/.claude/settings.json.
-    const subcommand = args[1] ?? 'status'
-    const { InstallCommands } = await import('../core/commands/install')
-    const cmd = new InstallCommands()
-    const mdMode = args.includes('--md')
-    let result: Awaited<ReturnType<typeof cmd.install>> | undefined
-    if (subcommand === 'install') result = await cmd.install(null, process.cwd(), { md: mdMode })
-    else if (subcommand === 'uninstall')
-      result = await cmd.uninstall(null, process.cwd(), { md: mdMode })
-    else {
-      const s = await cmd.status()
-      if (s.success) {
-        console.log(
-          mdMode
-            ? `# prjct Claude Code hooks\n\n- installed: ${s.installed}\n- expected: ${s.expected}\n`
-            : `installed: ${s.installed}/${s.expected}`
-        )
-        result = s
-      } else {
-        console.error(s.error)
-        result = s
-      }
-    }
-    process.exitCode = result.success ? 0 : 1
-  } else if (args[0] === 'hook') {
-    // `prjct hook <name>` — runs a single Claude Code hook. Invoked by
-    // the entries installed via `prjct claude install`. Reads stdin,
-    // writes JSON to stdout, exits 0 even on internal failure so the
-    // host session is never disturbed.
-    const hookName = args[1]
-    const projectPath = process.cwd()
-    try {
-      switch (hookName) {
-        case 'session-start': {
-          const { runSessionStartHook } = await import('../core/hooks/session-start')
-          await runSessionStartHook(projectPath)
-          break
-        }
-        case 'prompt': {
-          const { runPromptHook } = await import('../core/hooks/prompt')
-          await runPromptHook(projectPath)
-          break
-        }
-        case 'pre-commit': {
-          const { runPreCommitHook } = await import('../core/hooks/pre-commit')
-          await runPreCommitHook(projectPath)
-          break
-        }
-        case 'pre-edit': {
-          const { runPreEditHook } = await import('../core/hooks/pre-edit')
-          await runPreEditHook(projectPath)
-          break
-        }
-        case 'post-edit': {
-          const { runPostEditHook } = await import('../core/hooks/post-edit')
-          await runPostEditHook(projectPath)
-          break
-        }
-        case 'stop': {
-          const { runStopHook } = await import('../core/hooks/stop')
-          await runStopHook(projectPath)
-          break
-        }
-        case 'subagent-start': {
-          const { runSubagentStartHook } = await import('../core/hooks/subagent-start')
-          await runSubagentStartHook(projectPath)
-          break
-        }
-        case 'cwd-changed': {
-          const { runCwdChangedHook } = await import('../core/hooks/cwd-changed')
-          await runCwdChangedHook(projectPath)
-          break
-        }
-        default:
-          // Unknown hook: emit empty object, stay out of the way.
-          process.stdout.write('{}\n')
-      }
-      process.exitCode = 0
-    } catch {
-      process.stdout.write('{}\n')
-      process.exitCode = 0
-    }
-  } else if (args[0] === 'crew') {
-    // `prjct crew install|uninstall|status|record-run|checkpoints …`
-    // Manages the multi-agent crew bundle. Strictly opt-in.
-    const subcommand = args[1] ?? 'status'
-    const { CrewCommands } = await import('../core/commands/crew')
-    const cmd = new CrewCommands()
-    const mdMode = args.includes('--md')
-    const getFlag = (name: string): string | undefined => {
-      const idx = args.indexOf(`--${name}`)
-      if (idx >= 0 && idx + 1 < args.length && !args[idx + 1].startsWith('--')) {
-        return args[idx + 1]
-      }
-      return undefined
-    }
-    let result: { success: boolean; error?: string } = { success: false }
-    if (subcommand === 'install') {
-      result = await cmd.install(null, process.cwd(), { md: mdMode })
-    } else if (subcommand === 'uninstall') {
-      result = await cmd.uninstall(null, process.cwd(), { md: mdMode })
-    } else if (subcommand === 'status') {
-      result = await cmd.status(null, process.cwd(), { md: mdMode })
-    } else if (subcommand === 'checkpoints') {
-      // `prjct crew checkpoints [show|set|reset|export] [--content|--file]`
-      const checkpointsSub = args[2] ?? 'show'
-      result = await cmd.checkpoints(checkpointsSub, process.cwd(), {
-        md: mdMode,
-        content: getFlag('content'),
-        file: getFlag('file'),
-      })
-    } else if (subcommand === 'record-run') {
-      result = await cmd.recordRun(process.cwd(), {
-        md: mdMode,
-        spec: getFlag('spec'),
-        task: getFlag('task'),
-        'implementer-summary': getFlag('implementer-summary'),
-        files: getFlag('files'),
-        'reviewer-verdict': getFlag('reviewer-verdict'),
-        'reviewer-notes': getFlag('reviewer-notes'),
-        'run-id': getFlag('run-id'),
-      })
-    } else {
-      console.error(
-        `Unknown crew subcommand: ${subcommand}. Use: install, uninstall, status, checkpoints, record-run.`
-      )
-      result = { success: false, error: `unknown subcommand: ${subcommand}` }
-    }
-    process.exitCode = result.success ? 0 : 1
-  } else if (args[0] === 'doctor') {
-    const done = await trackSession('doctor')
-    const { doctorService } = await import('../core/services/doctor-service')
-    const exitCode = await doctorService.run(process.cwd())
-    process.exitCode = exitCode
-    done()
-  } else if (args[0] === 'uninstall') {
-    const { uninstall } = await import('../core/commands/uninstall')
-    const force = args.includes('--force') || args.includes('-f')
-    const backup = args.includes('--backup') || args.includes('-b')
-    const dryRun = args.includes('--dry-run') || args.includes('-n')
-    const keepPackage = args.includes('--keep-package')
-    const result = await uninstall({ force, backup, dryRun, keepPackage })
-    process.exitCode = result.success ? 0 : 1
-  } else if (args[0] === 'watch') {
-    const projectPath = process.cwd()
-    const projectId = await configManager.getProjectId(projectPath)
-
-    if (!projectId) {
-      console.error('No prjct project found. Run "prjct init" first.')
-      process.exitCode = 1
-    } else {
-      const { watchService } = await import('../core/services/watch-service')
-      const verbose = args.includes('--verbose') || args.includes('-v')
-      const debounceArg = args.find((a) => a.startsWith('--debounce='))
-      const debounceMs = debounceArg ? parseInt(debounceArg.split('=')[1], 10) : undefined
-      const intervalArg = args.find((a) => a.startsWith('--interval='))
-      const minIntervalMs = intervalArg ? parseInt(intervalArg.split('=')[1], 10) * 1000 : undefined
-
-      const result = await watchService.start(projectPath, {
-        verbose,
-        quiet: isQuietMode,
-        debounceMs,
-        minIntervalMs,
-      })
-
-      if (!result.success) {
-        console.error(result.error)
-        process.exitCode = 1
-      }
-    }
-  } else if (args[0] === 'help' || args[0] === '-h' || args[0] === '--help') {
-    const { getHelp } = await import('../core/utils/help')
-    const topic = args[1]
-    console.log(getHelp(topic))
-    process.exitCode = 0
-  } else if (args[0] === 'version' || args[0] === '-v' || args[0] === '--version') {
-    const detection = await detectAllProviders(isRefresh)
-    const home = os.homedir()
-    const cwd = process.cwd()
-    const [
-      claudeConfigured,
-      geminiConfigured,
-      cursorDetected,
-      cursorConfigured,
-      windsurfDetected,
-      windsurfConfigured,
-    ] = await Promise.all([
-      fileExists(path.join(home, '.claude', 'commands', 'p.md')),
-      fileExists(path.join(home, '.gemini', 'commands', 'p.toml')),
-      fileExists(path.join(cwd, '.cursor')),
-      fileExists(path.join(cwd, '.cursor', 'rules', 'prjct.mdc')),
-      fileExists(path.join(cwd, '.windsurf')),
-      fileExists(path.join(cwd, '.windsurf', 'rules', 'prjct.md')),
-    ])
-
-    const { providerStatusHeader, providerStatusLine } = await import(
-      '../core/utils/provider-status'
-    )
-    console.log(providerStatusHeader(VERSION))
-    console.log(
-      providerStatusLine(
-        'Claude Code',
-        detection.claude.installed ? (claudeConfigured ? 'ready' : 'installed') : 'missing',
-        detection.claude.version ? chalk.dim(` (v${detection.claude.version})`) : ''
-      )
-    )
-    console.log(
-      providerStatusLine(
-        'Gemini CLI',
-        detection.gemini.installed ? (geminiConfigured ? 'ready' : 'installed') : 'missing',
-        detection.gemini.version ? chalk.dim(` (v${detection.gemini.version})`) : ''
-      )
-    )
-    console.log(
-      providerStatusLine(
-        'Cursor IDE',
-        cursorDetected ? (cursorConfigured ? 'ready' : 'detected') : 'missing',
-        cursorDetected ? chalk.dim(' (project)') : '',
-        '○ not detected'
-      )
-    )
-    console.log(
-      providerStatusLine(
-        'Windsurf IDE',
-        windsurfDetected ? (windsurfConfigured ? 'ready' : 'detected') : 'missing',
-        windsurfDetected ? chalk.dim(' (project)') : '',
-        '○ not detected'
-      )
-    )
-
-    console.log(`
-${chalk.dim("Run 'prjct start' to configure (CLI providers)")}
-${chalk.dim("Run 'prjct init' to configure (Cursor/Windsurf IDE)")}
-${chalk.cyan('https://prjct.app')}
-`)
-  } else {
+  const { runBinCommand } = await import('../core/cli/bin-commands')
+  const handled = await runBinCommand(args, { isQuietMode, isRefresh, trackSession })
+  if (!handled) {
     // Default: check setup, auto-update, then run.
     // Route through pathManager so PRJCT_CLI_HOME is honored. Using
     // os.homedir() directly here bypassed the override — the same
